Mapping the Bb Content Item
This is a tentative map. I'm pretty sure there are a view other basic XML structures, but this the big one we need to start with.

Still to come (or at least to be assessed for worthiness):

NAVIGATIONAPLICATIONS
NAVIGATIONSETTINGS
GRADEBOOK
lom (for Content Collection files)
ANNOUNCEMENT
NOTIFICATIONRULES
questestinterop (tests and quizzes)
ASSESSMENTCREATIONSETTINGS
LINK (this might fix some of my frustration with course links!)
GROUPS
FORUM

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>