Routing Complete
All Blackboard items appear (so far, fingers crossed) that show up in the course Table of Contents in the manifest (so, not the other resources, like announcements and course links) seem to be being routed to the correct Canvas type. I've continued to update the documentation of that routing as I go.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>