File Attachments
Fixed the file uploading by opting not to try and use the POST call to upload the file, but instead transferring it through a public URL — after munging the filename to obscure it.

Subitems were not importing and indenting ([Issue 4](https://github.com/smtech/canvas/issues/4)) because they were files and files weren't even attempted. Now they are. Now it should be working.

Still to come, in order:

- Embedded links and images
- Assignment details
- Course links
- Quizzes
- Conferences

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>