Disambiguated SIS ID
Bah. So much for trying to be smart. If the user wants to create another course with the same SIS ID, I say "feh!". Instead, all SIS IDs are now a combination of the original course ID from Bb and a timestamp for the import. (Also: cleaned up timestamp formatting.)

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>