Attachments Redux
Reworked the attachment code and the content collection code to not conflict. (I think the core problem is that the PHP Zip class is not friendly towards windows paths, which is what Bb generated in their export files. Blergh.)

Started in on embedded links (simple ones might even be working).

Broadly speaking, this death star is operational. It still needs a few frills, but the hard part is done.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>