Embedded Links
Embedded Links appear to be working (at least, for files that exist in the ExportFile… it appears that Blackboard did not export _all_ content collection items with the course, so some courses may refer to files that simply aren't in the export. I think the Bb only exported things that were in the teacher's personal content collection or the course's content collection — but not things that were from other content collections. Maybe.)

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>