Announcements
Announcements are now imported (albeit without their date — Canvas won't let you pre-date an announcement). No effort is made to order them — although I suspect that they are in the manifest in the order they were made, so they should show up in Canvas in a similar or reverse order. The original post date has been appended to the title.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>