Argle-Bargle
Okay. I'm starting to feel [a teensy, tiny bit frustrated](https://twitter.com/battis/status/358749288710418433).

That said, all assignments are now being imported, with point values and due dates. (I think the so-called "Uncategorized" assignments _may_ correlate with assignments that were in the gradebook, but never linked from a Content Area. Maybe. (They also don't have matching resource files, which makes them harder to work with in my script, so…) It appears that Canvas disregards the position of an assignment in the list (or maybe it disregards the position if the number is greater than the total number of assignments currently in the list — so it doesn't remember positions for later, when there _are_ that many…). However, Bb seems to list the assignments in chronological order, so disregarding the position information appears harmless. (Of course, the chronology may be by creation date, not due date. Whatever.)

Embedded file attachments seem to be coming through fine, so long as they actually exist. There are a few cases where the ExportFiles refer to the X-ID of files that simply aren't in the export file — and I, honestly, cannot detect a pattern that explains how and when Bb decided to export or not export files from the Content Collection. Which is deeply troubling. And something we have no control over.

From a pragmatic standpoint, I think this is a working tool.

However, of course, there are a few more frills that I would like to add:
- Import Receipt.xml has received no attention in quite a while, and is now a fairly incomplete representation of the actual import. I would like to fix that!
- I would like to also upload a copy of the original ExportFile to the Canvas course, for posterity.
- I would like to import quizzes (there are very few in my test set, so it's not really feeling worth it to expend a lot of energy on it, but it would be, y'know… _complete_ to do that!)
- I would like to catch the few remaining missing embedded files and more successfully turn them into links to explanatory pages (as I did with file attachments). However, that would also entail identifying the context for the embed link — src or href — and thinking about how to handle those missing src files…
- I would like to tidy up the code and clean up the documentation.
- I would like a cookie.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>