Import Receipt
Lots of restructuring of code, and some changes of naming conventions.

- $___Xml refers to a SimpleXMLElement, while $____Dom refers to a DOMElement — there were getting to be enough of each that it made sense to start to identify them explicitly.
- $xpathResult is always where an XPath search is stored, as a reminder that it's always an array and should be treated as such.
- $item and $res are now $itemXml and $resXml (as noted above), for clarity throughout.
- Most (all?) extraneous parameters have been extirpated (they had been a good idea in the initial design, but things change).

Import Receipts are now generated completely (oh, so _very_ completely!), with all of the import information stored in the canvas namespace. (It might be nice to filter out things that are also going into CDATA nodes, but… what the hey). This was suddenly an urgent priority, when I realized that course links really needed that complete a record to work from to rebuild the links. And once course links are done, this puppy can be put to bed!

I think I've been seeing a lack of responsiveness in the Canvas Test instance, but I may also have an infinite loop somewhere… unlikely though that may be.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>