Archive Discussions
A first, quick pass at a tool to archive discussions. Initially meant just to archive discussions out of a group, for potential reimport into a course, it's written so that it should be able to archive discussions out of a course, group or collection.

N.B. Because the group that I was archiving had less than 50 discussions, I didn't bother to deal with the pagination of Canvas' JSON responses proactively!

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>