DisplayError and Debugging
Updated displayError to be DEBUGGING-sensitive, allowing yours truly to flag some displayError calls with a debugging level, for easier toggling on and off of debugging messages.

Also standardized on '%' as my regexp delimiter across the board (since '|' is useful as an OR operator, which, for whatever reason, I hadn't needed until now).

Fixed a typo in the displayPage footer.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>