Syncing
Hourly, daily and weekly syncing are now available options. Choosing one of these options will schedule a crontab job (if it does not already exist) to trigger the scheduled import (via sync.php).

There is room in here to update the form to create custom schedules as well, but I haven't bothered with that yet.

OAuth is becoming increasingly… pertinent (as is the DELETEing of calendar events)...

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>