Sync Fix
Realized that, for the PHP includes to work, the crontab commands would need to come from within our script directory. Added shell script sync.sh that resolves this and tweaked the sync scheduling to call sync.sh instead of 'php sync.php'.

Added in an attempt to ensure that sync.sh is executable (probably will be, if normal web directory permissions of 0775 are set, I hope…).

Added some clean-up to the scheduler, so that unused crontabs are deleted.

Also, if you leave me with too much time to fiddle around with the debugger, I'll go overboard… and now I've got a nicer way of identifying which instance of the script is posting to the log file. Doubtless, I'll keep fiddling with this indefinitely.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>