Safer Function Naming Pattern
I realized that just tossing my arbitrarily named Javascript functions into the same namespace as the app that I'm modifying… could end badly.

Now all of my functions are named stmarks_*.

'would be good to move my variables out of that namespace as well, but so far nothing is breaking!

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>