Delays no longer arbitrary
Gee, if the problem is that the code needs a particular element or variable to be ready before it runs, howsabout not just guessing how long it _might_ take, but just waiting for it to be ready?

Gosh, Seth, that's a good idea!

Darn right it is. So now the code will simply keep retrying until the elements it needs for its tests are ready and can be used.

For dumb.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>