Dynamically Load Students
Switched over to using AJAX to load student menu, so that faculty will see the menu, even before it's loaded.

(Also: vertically centered everything so that I would stop getting heebie-jeebies. Opted not to force DIVs to be TABLEs in my HTML/CSS, which means I have a bunch of old-school TABLEs. Sigh.)

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>