Reusable Calculations
Pulled calculations that need to be reused in summaries out of graphs into common.inc.php, so that the same calculations are used for both numeric and visual summaries. It means a few more MySQL queries and calculations overall, but they're fast and we'll live with it, if it gets us consistent (and hopefully accurate) results.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>