Fine-tuning Module Item Indent
I think this works pretty well across all browsers tested (Chrome, Safari, Firefox on Mac). The indent for each level is the height of the row (so... a nice square grid). Forcing the type icons to be consistently visible — and removing their extraneous padding-right — allows all types of content to be indented the same amount.

Oy.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>