New Header
The new header background image from www.stmarksschool.org. We should
also shift to the lion, instead of the crest (although we will not use
the wide version of the school name with Southborough beneath, for
space reasons).

Would also be worth exploring how to best make the login page connect
back to the overall external branding.