Embed Google Drive Search
A stand-alone form to embed in a wiki page (or anywhere else on the web) to search Google Drive.

Lots of GET parameters to configure it.

Nota been: the "within" parameter is a touch misleading. In the change from Google Docs to Google Drive, Google broke/lost/misplaced the "in:" keyword to restrict results to a specific folder. However, you _can_ restrict searches to files shared to a specific user with the "to:" keyword. So, for my purposes, I have a unique user with whom I have shared the folder that I want to search within specifically on the drive, so my within parameter would be:

within=to:uniqueuser@gmail.com

(Or whatever the unique user's email address is.)

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>