Summer Course Merge (part 1)
Mostly done with the script. Just need to decide how to handle the courses that will be deleted. (I think... put them in the Sandbox and send a message to the users enrolled as teachers asking what to do.)

Still need to test the 'modify' action, but I think it's all good to go.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>