CSV Import and Clone
A script to take a standard SIS CSV import of courses, with the addition of a template_id column, and create those courses as clones of the template course. The template_id has to be given as a Canvas ID (not an SIS ID) for the template course. Currently, it seems to work okay, with the exception that navigation settings don't carry over from the template to the clone (since there seems to be no API access to the navigation settings, or maybe I'm just missing it).