Template relative paths
This will make it much easier to re-organize/rename scripts! (I hope.) It feels a little hacky, but seems to work fine. The trick is just that we have to query `$smarty` for the `current_dir` before we enter our `extends` clause, as we are (once in the `extends` clause) then evaluating the template being extended, rather than the extending template. (I, for one, think that doesn't quite make sense. But wharves.)