Unfavorite Hidden Courses
Use the Custom Prefs settings to intelligently "hide" hidden courses from users. The complication is that, as is the case for many users, they may not have already set their own favorites explicitly. This means that all of their active courses are returned implicitly as favorites. However, implicit favorites cannot be removed. So... the fix is to take the current favorites list, whether it is implicit or explicit, and walk through it and (re)set favorites that are _not_ on the list of hidden courses, while deleting favorites that _are_ on the hidden favorites list. This prevents users who have set their favorites from being adversely impacted, while preserving the experience for users who have not set their favorites. I hope.