Update Faculty Account Admin
A script to support our "radically open" Canvas structure: faculty are account admins within their sub-accounts, able to see all of the other courses (not grades or student work, but what other teachers post). This lets you pick a sub-account and terms in which to make the teachers of the courses (in that sub-account and terms) a particular role within the sub-account.